+++ src/test/java/com/petproject/ecommerce/generators/ProductDataGenerator.java
@@ -2,6 +2,8 @@ package com.petproject.ecommerce.generators;
 
 import com.github.javafaker.Faker;
 
+import java.math.BigDecimal;
+
 public class ProductDataGenerator {
     private static final Faker FAKER = new Faker();
 
@@ -9,7 +11,7 @@ public class ProductDataGenerator {
         return FAKER.commerce().productName();
     }
 
-    public static Double randomPrice() {
-        return FAKER.number().randomDouble(2, 1, 10000);
+    public static BigDecimal randomPrice() {
+        return BigDecimal.valueOf(FAKER.number().randomDouble(2, 1, 10000));
     }
 }
